// uni_modules/uview-pro/types/global.d.ts
export type FormRuleItem = {
    required?: boolean;
    message?: string;
    trigger?: string | string[];
    min?: number;
    max?: number;
    pattern?: RegExp;
    type?: string;
    validator?: (rule: any, value: any, callback: any) => boolean;
    asyncValidator?: (rule: any, value: any, callback: any) => void;
    fields?: FormRules;
    defaultField?: FormRuleItem;
};
export type FormRules = Record<string, FormRuleItem | FormRuleItem[]>;

export type InputType =
    | 'text'
    | 'number'
    | 'idcard'
    | 'digit'
    | 'password'
    | 'textarea'
    | 'phone'
    | 'url'
    | 'email'
    | 'safe-password'
    | 'name'
    | 'bank-card'
    | 'tel'
    | 'select';

export type InputAlign = 'left' | 'center' | 'right';

export type InputConfirmType = 'send' | 'search' | 'next' | 'go' | 'done';

export type InputLabelPosition = 'left' | 'top';

export type TextareaBorder = 'surround' | 'none' | 'bottom';

export type FormErrorType = 'message' | 'border' | 'border-bottom' | 'none' | 'toast';